rename project and library back to odo because open source and pip
